requirements.txt
amqp==2.2.2
anyjson==0.3.3
beautifulsoup4==4.6.0
billiard==3.5.0.3
celery==4.1.0
certifi==2018.1.18
chardet==3.0.4
cssselect==1.0.3
Django==2.0
django-celery==3.2.2
django-celery-beat==1.1.1
django-celery-results==1.0.1
django-ckeditor==5.4.0
django-filter==1.1.0
django-js-asset==1.0.0
djangorestframework==3.7.7
eventlet==0.22.1
greenlet==0.4.13
idna==2.6
kombu==4.1.0
lxml==4.1.1
mysqlclient==1.3.12
Pillow==5.0.0
pytz==2018.3
redis==2.10.6
reportlab==3.4.0
requests==2.18.4
urllib3==1.22
vine==1.1.4
